% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corarrange.R
\name{corarrange}
\alias{corarrange}
\title{Arrange numeric variables by correlation.}
\usage{
corarrange(dataframe, var)
}
\arguments{
\item{var}{(Optional) filter by specific variable}

\item{matrix}{Dataframe, numeric values}
}
\value{
Table of correlations, arranged by correlation coefficient
}
\description{
Arrange numeric variables by correlation.
}
