% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsize.R
\name{effectsize}
\alias{effectsize}
\title{Table of effect size}
\usage{
effectsize(formula, data)
}
\arguments{
\item{formula}{y~x. y numeric, x categorical, 2 group only}

\item{data}{Dataframe}
}
\value{
Table of effect size
}
\description{
Table of effect size
}
