% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_histogram.R
\name{sd_histogram}
\alias{sd_histogram}
\title{Sampling distribution histogram}
\usage{
sd_histogram(formula, data, n, alpha = 0.05, statistic = b1, region = middle)
}
\arguments{
\item{formula}{y~x. y numeric, x numeric}

\item{data}{dataframe}

\item{n}{sample size}

\item{alpha}{alpha value (0.05)}

\item{statistic}{b1, pre, f, etc.}

\item{region}{middle, lower, upper}
}
\description{
Sampling distribution histogram
}
