% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corarrange.R
\name{corarrange}
\alias{corarrange}
\title{Arranged cor matrix}
\usage{
corarrange(matrix, var)
}
\arguments{
\item{matrix}{Matrix}

\item{var}{(Optional) filter by specific variable}
}
\value{
Arranged cor matrix
}
\description{
Arrange a cor matrix by correlation.
}
